Actualizo el nombre del game model para no tener que andar cambiando
