+++ TGC.Group/Model/GameModel.cs
@@ -1,5 +1,4 @@
 using Microsoft.DirectX.DirectInput;
-using System;
 using System.Drawing;
 using TGC.Core.Direct3D;
 using TGC.Core.Example;
@@ -7,7 +6,13 @@ using TGC.Core.Input;
 using TGC.Core.Mathematica;
 using TGC.Core.SceneLoader;
 using TGC.Core.SkeletalAnimation;
+using TGC.Core.Geometry;
+using TGC.Core.Mathematica;
+using TGC.Core.BoundingVolumes;
+using TGC.Core.Collision;
+using TGC.Core.Textures;
 using TGC.Group.Camera;
+using System;
 
 namespace TGC.Group.Model
 {
@@ -34,11 +39,32 @@ namespace TGC.Group.Model
         //Boleano para ver si dibujamos el boundingbox
         private bool BoundingBox { get; set; }
         private const float VELOCIDAD_DESPLAZAMIENTO = 50f;
-        private TgcScene scene;
+        private TgcScene escenaPlataformas;
         private TgcSkeletalMesh personaje;
         private GameCamera camara;
         private TGCVector3 movimiento;
         private TGCMatrix ultimaPos;
+
+        private bool estaEnLaPlataforma1;
+        private bool estaEnLaPlataforma2;
+
+        // Planos de limite
+        private TgcMesh planoIzq;
+        private TgcMesh planoDer;
+        private TgcMesh planoFront;
+        private TgcMesh planoBack;
+
+        // Plataformas
+        private TgcMesh plataforma1;
+        private TgcMesh plataforma2;
+
+        //Transformaciones de plataforma
+        private TGCMatrix transformacionBox;
+        private TGCMatrix transformacionBox2;
+
+        //Constantes para velocidades de movimiento de plataforma
+        private const float MOVEMENT_SPEED = 1f;
+        private float orbitaDeRotacion;
 
         /// <summary>
         ///     Se llama una sola vez, al principio cuando se ejecuta el ejemplo.
@@ -50,10 +76,35 @@ namespace TGC.Group.Model
         {
             //Device de DirectX para crear primitivas.
             var d3dDevice = D3DDevice.Instance.Device;
-            
+
             var loader = new TgcSceneLoader();
-            scene = loader.loadSceneFromFile(MediaDir + "primer-nivel\\pozo-plataformas\\tgc-scene\\plataformas\\plataformas-TgcScene.xml");
-            
+            escenaPlataformas = loader.loadSceneFromFile(MediaDir + "primer-nivel\\pozo-plataformas\\tgc-scene\\plataformas\\escenarioPlataformas-TgcScene.xml");
+
+            plataforma1 = loader.loadSceneFromFile(MediaDir + "primer-nivel\\pozo-plataformas\\tgc-scene\\plataformas\\plataforma-TgcScene.xml").Meshes[0];
+
+            plataforma2 = plataforma1.createMeshInstance(plataforma1.Name + "2");
+
+            plataforma1.AutoTransform = false;
+            plataforma2.AutoTransform = false;
+
+            planoIzq = loader.loadSceneFromFile(MediaDir + "primer-nivel\\pozo-plataformas\\tgc-scene\\plataformas\\planoHorizontal-TgcScene.xml").Meshes[0];
+            planoIzq.AutoTransform = false;
+
+            planoDer = planoIzq.createMeshInstance("planoDer");
+            planoDer.AutoTransform = false;
+            planoDer.Transform = TGCMatrix.Translation(-39, 0, 0);
+            planoDer.BoundingBox.transform(planoDer.Transform);
+
+            planoFront = loader.loadSceneFromFile(MediaDir + "primer-nivel\\pozo-plataformas\\tgc-scene\\plataformas\\planoVertical-TgcScene.xml").Meshes[0];
+            planoFront.AutoTransform = false;
+            planoFront.Transform = TGCMatrix.Translation(50, 0, -65);
+            planoFront.BoundingBox.transform(planoFront.Transform);
+
+            planoBack = planoFront.createMeshInstance("planoBack");
+            planoBack.AutoTransform = false;
+            planoBack.Transform = TGCMatrix.Translation(0, 0, 135);
+            planoBack.BoundingBox.transform(planoBack.Transform);
+
 
             var skeletalLoader = new TgcSkeletalLoader();
             personaje = skeletalLoader.loadMeshAndAnimationsFromFile(
@@ -65,9 +116,7 @@ namespace TGC.Group.Model
                     MediaDir + "primer-nivel\\pozo-plataformas\\tgc-scene\\Robot\\Parado-TgcSkeletalAnim.xml"
                 });
 
-
             personaje.AutoTransform = false;
-            personaje.Transform = TGCMatrix.Identity;
 
             //Configurar animacion inicial
             personaje.playAnimation("Parado", true);
@@ -77,14 +126,7 @@ namespace TGC.Group.Model
             ultimaPos = TGCMatrix.Translation(personaje.Position);
 
             camara = new GameCamera(personaje.Position, 100, 200);
-            //var cameraPosition = new TGCVector3(0, 0, 200);
-            //Quiero que la camara mire hacia el origen (0,0,0).
-            //var lookAt = TGCVector3.Empty;
-            //Configuro donde esta la posicion de la camara y hacia donde mira.
-            //Camara.SetCamera(cameraPosition, lookAt);
             Camara = camara;
-
-            BoundingBox = true;
 
         }
 
@@ -96,6 +138,8 @@ namespace TGC.Group.Model
         public override void Update()
         {
             PreUpdate();
+
+            AnimarPlataformas();
 
             CalcularMovimiento();
 
@@ -109,52 +153,6 @@ namespace TGC.Group.Model
             PostUpdate();
         }
 
-        private void CalcularMovimiento()
-        {
-            var velocidadCaminar = VELOCIDAD_DESPLAZAMIENTO * ElapsedTime;
-
-            //Calcular proxima posicion de personaje segun Input
-            var moving = false;
-            movimiento = TGCVector3.Empty;
-            //transformacionPersonaje = TGCMatrix.Identity;
-
-            if (Input.keyDown(Key.W))
-            {
-                movimiento += new TGCVector3(0, 0, -velocidadCaminar);
-                moving = true;
-            }
-
-            if (Input.keyDown(Key.S))
-            {
-                movimiento += new TGCVector3(0, 0, velocidadCaminar);
-                moving = true;
-            }
-
-            if (Input.keyDown(Key.D))
-            {
-                movimiento += new TGCVector3(-velocidadCaminar, 0, 0);
-                moving = true;
-            }
-
-            if (Input.keyDown(Key.A))
-            {
-                movimiento += new TGCVector3(velocidadCaminar, 0, 0);
-                moving = true;
-            }
-
-            if (moving)
-            {
-                personaje.playAnimation("Caminando", true);
-
-                //personaje.Position += movimiento;
-                ultimaPos *= TGCMatrix.Translation(movimiento);
-            }
-            else
-            {
-                personaje.playAnimation("Parado", true);
-            }
-        }
-
         /// <summary>
         ///     Se llama cada vez que hay que refrescar la pantalla.
         ///     Escribir aquí todo el código referido al renderizado.
@@ -165,45 +163,240 @@ namespace TGC.Group.Model
             //Inicio el render de la escena, para ejemplos simples. Cuando tenemos postprocesado o shaders es mejor realizar las operaciones según nuestra conveniencia.
             PreRender();
 
-            personaje.Transform = 
-                TGCMatrix.Scaling(personaje.Scale)
-                * TGCMatrix.RotationYawPitchRoll(personaje.Rotation.Y, personaje.Rotation.X, personaje.Rotation.Z)
-                * ultimaPos;
+            RenderizarPlataformas();
+
+            if (estaEnLaPlataforma1)
+            {
+                personaje.Transform = transformacionBox;
+            }
+            if (estaEnLaPlataforma2)
+            {
+                personaje.Transform = TGCMatrix.Scaling(personaje.Scale) * transformacionBox2;
+
+               // personaje.Transform =
+               //TGCMatrix.Scaling(personaje.Scale)
+               //            * TGCMatrix.RotationYawPitchRoll(personaje.Rotation.Y, personaje.Rotation.X, personaje.Rotation.Z)
+               //            * TGCMatrix.Translation(personaje.Position)
+               //            * transformacionBox2;
+            }
+            else
+            {
+                personaje.Transform =
+           TGCMatrix.Scaling(personaje.Scale)
+                       * TGCMatrix.RotationYawPitchRoll(personaje.Rotation.Y, personaje.Rotation.X, personaje.Rotation.Z)
+                       * TGCMatrix.Translation(personaje.Position)
+                       * ultimaPos;
 
             personaje.BoundingBox.transform(personaje.Transform);
+            }
+
+            personaje.BoundingBox.Render();
 
             personaje.animateAndRender(ElapsedTime);
 
-            Console.WriteLine("nombre: {0} \n", personaje.Name);
-
-            scene.RenderAll();
-
-            //Siempre antes de renderizar el modelo necesitamos actualizar la matriz de transformacion.
-            //Debemos recordar el orden en cual debemos multiplicar las matrices, en caso de tener modelos jerárquicos, tenemos control total.
-            //Box.Transform = TGCMatrix.Scaling(Box.Scale) * TGCMatrix.RotationYawPitchRoll(Box.Rotation.Y, Box.Rotation.X, Box.Rotation.Z) * TGCMatrix.Translation(Box.Position);
-
-
-            //A modo ejemplo realizamos toda las multiplicaciones, pero aquí solo nos hacia falta la traslación.
-            //Finalmente invocamos al render de la caja
-            //Box.Render();
-
-            //Cuando tenemos modelos mesh podemos utilizar un método que hace la matriz de transformación estándar.
-            //Es útil cuando tenemos transformaciones simples, pero OJO cuando tenemos transformaciones jerárquicas o complicadas.
-            //Mesh.UpdateMeshTransform();
-            //Render del mesh
-            //Mesh.Render();
-
-            //Render de BoundingBox, muy útil para debug de colisiones.
             if (BoundingBox)
             {
+                planoBack.BoundingBox.Render();
+                planoFront.BoundingBox.Render();
+                planoIzq.BoundingBox.Render();
+                planoDer.BoundingBox.Render();
+                plataforma1.BoundingBox.Render();
+                plataforma2.BoundingBox.Render();
                 personaje.BoundingBox.Render();
-                foreach (var mesh in scene.Meshes) {
-                    mesh.BoundingBox.Render();
-                }
             }
+  
+            escenaPlataformas.RenderAll();
 
+   
             //Finaliza el render y presenta en pantalla, al igual que el preRender se debe para casos puntuales es mejor utilizar a mano las operaciones de EndScene y PresentScene
             PostRender();
+        }
+
+        private void AnimarPlataformas() {
+            //Muevo las plataformas
+            var Mover = TGCMatrix.Translation(0, 0, -10);
+            var Mover2 = TGCMatrix.Translation(0, 0, 65);
+
+            //Punto por donde va a rotar
+            var Trasladar = TGCMatrix.Translation(0, 0, 10);
+            var Trasladar2 = TGCMatrix.Translation(0, 0, -10);
+
+            //Aplico la rotacion
+            var Rot = TGCMatrix.RotationX(orbitaDeRotacion);
+
+            //Giro para que la caja quede derecha
+            var RotInversa = TGCMatrix.RotationX(-orbitaDeRotacion);
+
+            transformacionBox = Mover * Trasladar * Rot * Trasladar * RotInversa;
+            transformacionBox2 = Mover2 * Trasladar2 * RotInversa * Trasladar2 * Rot;
+        }
+
+        private void RenderizarPlataformas() {
+            //Dibujar la primera plataforma en pantalla
+            plataforma1.Transform = transformacionBox;
+            plataforma1.Render();
+            plataforma1.BoundingBox.transform(plataforma1.Transform);
+            plataforma1.BoundingBox.Render();
+
+            //Dibujar la segunda plataforma en pantalla
+            plataforma2.Transform = transformacionBox2;
+            plataforma2.Render();
+            plataforma2.BoundingBox.transform(plataforma2.Transform);
+            plataforma2.BoundingBox.Render();
+
+            //Recalculamos la orbita de rotacion
+            orbitaDeRotacion += MOVEMENT_SPEED * ElapsedTime;
+        }
+
+        private void CalcularMovimiento()
+        {
+            var velocidadCaminar = VELOCIDAD_DESPLAZAMIENTO * ElapsedTime;
+            estaEnLaPlataforma1 = false;
+            estaEnLaPlataforma2 = false;
+
+            //Calcular proxima posicion de personaje segun Input
+            var moving = false;
+            movimiento = TGCVector3.Empty;
+            
+            if (Input.keyDown(Key.W))
+            {
+                movimiento.Z = -velocidadCaminar;
+                moving = true;
+            }
+
+            if (Input.keyDown(Key.S))
+            {
+                movimiento.Z = velocidadCaminar;
+                moving = true;
+            }
+
+            if (Input.keyDown(Key.D))
+            {
+                movimiento.X = -velocidadCaminar;
+                moving = true;
+            }
+
+            if (Input.keyDown(Key.A))
+            {
+                movimiento.X = velocidadCaminar;
+                moving = true;
+            }
+
+            if (ChocoConLimite(personaje, planoIzq))
+            {
+                NoMoverHacia(Key.A);
+
+            }
+            else if (ChocoConLimite(personaje, planoDer))
+            {
+                NoMoverHacia(Key.D);
+            }
+            else { // no hay colisiones contra los planos laterales
+                ultimaPos *= TGCMatrix.Translation(movimiento);
+            }
+
+            if (ChocoConLimite(personaje, planoFront))
+            { // HUBO CAMBIO DE ESCENARIO
+                /* Aca deberiamos hacer algo como no testear mas contra las cosas del escenario anterior y testear
+                  contra las del escenario actual. 
+                */
+
+                planoFront.BoundingBox.setRenderColor(Color.AliceBlue);
+            }
+            else {
+                planoFront.BoundingBox.setRenderColor(Color.Yellow);
+            }
+
+            if (ChocoConLimite(personaje, planoBack))
+            { // HUBO CAMBIO DE ESCENARIO
+              /* Aca deberiamos hacer algo como no testear mas contra las cosas del escenario anterior y testear
+                  contra las del escenario actual. 
+              */
+
+                planoBack.BoundingBox.setRenderColor(Color.AliceBlue);
+            }
+            else {
+                planoBack.BoundingBox.setRenderColor(Color.Yellow);
+            }
+
+            if (ChocoConPlataforma(personaje, plataforma1))
+            {
+                estaEnLaPlataforma1 = true;
+
+                //MoverSegunPlataforma(personaje, plataforma1);
+            }
+            else
+            {
+                //ultimaPos *= TGCMatrix.Translation(movimiento);
+            }
+
+            if (ChocoConPlataforma(personaje, plataforma2))
+            {
+                estaEnLaPlataforma2 = true;
+
+                //MoverSegunPlataforma(personaje, plataforma2);
+            }
+            else
+            {
+                //ultimaPos *= TGCMatrix.Translation(movimiento);
+            }
+
+            if (moving)
+            {
+                personaje.playAnimation("Caminando", true);
+            }
+            else
+            {
+                personaje.playAnimation("Parado", true);
+            }
+
+            camara.Target = personaje.Position;
+        }
+
+        private void MoverSegunPlataforma(TgcSkeletalMesh personaje, TgcMesh plataforma)
+        {
+            //ultimaPos *= transformacionBox;
+            //personaje.BoundingBox.transform(plataforma.Transform);
+        }
+
+        private void NoMoverHacia(Key key) {
+            switch(key)
+            {
+                case Key.A:
+                    if (movimiento.X > 0) // los ejes estan al reves de como pensaba, o lo entendi mal.
+                        ultimaPos *= TGCMatrix.Translation(0, movimiento.Y, movimiento.Z);
+                    else
+                        ultimaPos *= TGCMatrix.Translation(movimiento);
+                    break;
+                case Key.D:
+                    if (movimiento.X < 0)
+                        ultimaPos *= TGCMatrix.Translation(0, movimiento.Y, movimiento.Z);
+                    else
+                        ultimaPos *= TGCMatrix.Translation(movimiento);
+                    break;
+                case Key.S:
+                    if (movimiento.Z > 0)
+                        ultimaPos *= TGCMatrix.Translation(movimiento.X, movimiento.Y, 0);
+                    else
+                        ultimaPos *= TGCMatrix.Translation(movimiento);
+                    break;
+                case Key.W:
+                    if (movimiento.Z < 0)
+                        ultimaPos *= TGCMatrix.Translation(movimiento.X, movimiento.Y, 0);
+                    else
+                        ultimaPos *= TGCMatrix.Translation(movimiento);
+                    break;
+            }
+                
+        }
+
+        private bool ChocoConPlataforma(TgcSkeletalMesh personaje, TgcMesh plataforma)
+        {
+            return TgcCollisionUtils.testAABBAABB(plataforma.BoundingBox, personaje.BoundingBox);
+        }
+
+        private bool ChocoConLimite(TgcSkeletalMesh personaje, TgcMesh planoIzq) {
+            return TgcCollisionUtils.testAABBAABB(planoIzq.BoundingBox, personaje.BoundingBox);
         }
 
         /// <summary>
@@ -214,8 +407,11 @@ namespace TGC.Group.Model
         public override void Dispose()
         {
             //Dispose del mesh.
-            scene.DisposeAll();
+            escenaPlataformas.DisposeAll();
             personaje.Dispose();
+            planoIzq.Dispose(); // solo se borran los originales
+            planoFront.Dispose(); // solo se borran los originales
+            plataforma1.Dispose(); // solo se borran los originales
         }
     }
 }
